+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     with RSA or X25519 it still needs to encrypt a symmetric key to encrypt final content           
 * es: update to 7.6.2
 * db: added max operations per action check (default is 5000), to protect bad impl (e.g. infinite loop with db calls) or bad practice (make action too long for zero-down time release) 
+* db: added errorType (INTEGRITY_CONSTRAINT_VIOLATION) in UncheckedSQLException, to support catch duplicate key exception in business logic 
 
 ### 7.3.12.1 (3/25/2020 - 4/1/2020)
 * db: added query.in(field, params) shortcut to build dynamic "where in clause"